+++ src/core/settings.h
@@ -95,7 +95,7 @@ struct Values {
     // Renderer
     bool use_hw_renderer;
     bool use_shader_jit;
-    float resolution_factor;
+    u16 resolution_factor;
     bool use_vsync;
     bool use_frame_limit;
     u16 frame_limit;
